Extract Manager, Pool and Object into separate modules
